# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Batoi Bolsa  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/company": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Obtener lista de empresas",
                "description": "Devuelve una lista de empresas paginada",
                "operationId": "getCompanyIndex",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pagination": {
                                            "$ref": "#/components/schemas/CompanyPagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Crear una nueva empresa",
                "description": "Crea una nueva empresa con los datos proporcionados",
                "operationId": "createCompany",
                "requestBody": {
                    "description": "Datos de la empresa a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada no válidos"
                    }
                }
            }
        },
        "/api/company/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Obtener detalles de una empresa",
                "description": "Devuelve los detalles de una empresa específica",
                "operationId": "getCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "Actualizar una empresa",
                "description": "Actualiza una empresa existente con los datos proporcionados",
                "operationId": "updateCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos de la empresa a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            }
        },
        "/api/company/{cif}": {
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "Eliminar una empresa",
                "description": "Elimina una empresa por su CIF",
                "operationId": "deleteCompany",
                "parameters": [
                    {
                        "name": "cif",
                        "in": "path",
                        "description": "CIF de la empresa a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La empresa ha sido borrada con éxito"
                                        },
                                        "data": {
                                            "type": "integer",
                                            "example": 123
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            }
        },
...
        "/api/checkCIF/{CIF}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Verificar CIF de la empresa",
                "description": "Verifica si el CIF proporcionado pertenece a alguna empresa registrada",
                "operationId": "checkCIF",
                "parameters": [
                    {
                        "name": "CIF",
                        "in": "path",
                        "description": "Número de identificación fiscal (CIF) de la empresa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Indica si el CIF pertenece a una empresa registrada (true) o no (false)",
                                    "type": "boolean"
                                },
                                "example": true
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            }
        },
        "/api/companyCIF/{CIF}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Obtener empresa por CIF",
                "description": "Obtiene los detalles de una empresa basándose en su CIF",
                "operationId": "getCompanyCIF",
                "parameters": [
                    {
                        "name": "CIF",
                        "in": "path",
                        "description": "CIF (Código de Identificación Fiscal) de la empresa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Empresa no encontrada para el CIF proporcionado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompanyPagination": {
                "title": "Company Pagination",
                "description": "Esquema para la colección paginada de empresas",
                "properties": {
                    "current_page": {
                        "description": "Número de la página actual",
                        "type": "integer",
                        "example": 1
                    },
                    "total_pages": {
                        "description": "Número total de páginas",
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "description": "Número de elementos por página",
                        "type": "integer",
                        "example": 10
                    },
                    "total_records": {
                        "description": "Número total de registros",
                        "type": "integer",
                        "example": 47
                    },
                    "data": {
                        "description": "Datos de las empresas",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Company"
                        }
                    },
                    "status": {
                        "description": "Estado de la respuesta",
                        "type": "string",
                        "example": "success"
                    },
                    "links": {
                        "properties": {
                            "prev": {
                                "description": "Enlace a la página anterior",
                                "type": "string",
                                "example": "http://example.com/api/company?page=1"
                            },
                            "next": {
                                "description": "Enlace a la página siguiente",
                                "type": "string",
                                "example": "http://example.com/api/company?page=3"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    }
}